+++ packages/components-angular/src/app/examples/input-search-example-controlled.component.ts
@@ -1,0 +1,24 @@
+import { CUSTOM_ELEMENTS_SCHEMA, ChangeDetectionStrategy, Component } from '@angular/core';
+import { InputSearchInputEventDetail, PorscheDesignSystemModule } from '@porsche-design-system/components-angular';
+
+@Component({
+  selector: 'page-input-search-example-controlled',
+  template: `
+    <p-input-search name="some-name" label="Some Label" [value]="value" (input)="onInput($event)" />
+    <p-text>{{ debugText }}</p-text>
+  `,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [PorscheDesignSystemModule],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+})
+export class InputSearchExampleControlledComponent {
+  value: string = '';
+  get debugText(): string {
+    return `Value: ${this.value}`;
+  }
+
+  onInput(e: CustomEvent<InputSearchInputEventDetail>) {
+    this.value = (e.detail.target as HTMLInputElement).value;
+  }
+}