+++ packages/components/scripts/generateDSRComponents.ts
@@ -67,6 +67,7 @@ const generateDSRComponents = (): void => {
         .replace(/\s+onKeyDown={.*?}/g, '') // onKeyDown props
         .replace(/\s+onPaste={.*?}/g, '') // onPaste props
         .replace(/\s+onInput={.*?}/g, '') // onInput props
+        .replace(/\s+onWheel={.*?}/g, '') // onWheel props
         .replace(/\s+on(?:Tab)?Change={.*?}/g, '') // onChange and onTabChange props
         .replace(/\s+onUpdate={.*?}/g, '') // onUpdate props
         .replace(/ +ref: [\s\S]*?,\n/g, '') // ref props
@@ -224,9 +225,9 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
             .replace(/onBlur=\{onBlur}/g, '')
             .replace(/maxlength/, 'maxLength')
             .replace(/minlength/, 'minLength')
-            .replace(/readonly/, 'readOnly')
+            .replace(/\sreadonly/, 'readOnly')
             .replace(/autocomplete/, 'autoComplete')
-            .replace(/\b(onInput|onChange|onBlur|refElement\s*,?)/g, '// $1')
+            .replace(/\b(onInput|onWheel|onChange|onBlur|refElement\s*,?)/g, '// $1')
             .replace(
               /}\) => \{/,
               `$&
@@ -758,7 +759,7 @@ $&`
           .replace(/this\.props\.value = state;/, '')
           .replace(/formDisabledCallback\(disabled: boolean\)/, 'formDisabledCallback()')
           .replace(/formStateRestoreCallback\(state: string\)/, 'formStateRestoreCallback()');
-      } else if (tagName === 'p-input-number') {
+      } else if (tagName === 'p-input-number' || tagName === 'p-input-search') {
         newFileContent = newFileContent
           .replace(/@AttachInternals\(\)/, '')
           .replace(