+++ packages/components/src/components/common/input-base/input-base.tsx
@@ -1,7 +1,9 @@
 import { type FunctionalComponent, type JSX, h } from '@stencil/core';
 import type { Theme } from '../../../types';
+import { getPrefixedTagNames } from '../../../utils';
 import { Label } from '../label/label';
 import { descriptionId } from '../label/label-utils';
+import { LoadingMessage, loadingId } from '../loading-message/loading-message';
 import { StateMessage, messageId } from '../state-message/state-message';
 import type {
   InputBaseBlurEventDetail,
@@ -17,6 +19,8 @@ type InputBaseProps = {
   id: string;
   label?: string;
   description?: string;
+  loading?: boolean;
+  initialLoading?: boolean;
   required?: boolean;
   disabled?: boolean;
   state?: InputBaseState;
@@ -48,6 +52,8 @@ export const InputBase: FunctionalComponent<InputBaseProps> = ({
   id,
   label,
   description,
+  loading,
+  initialLoading,
   required,
   disabled,
   state,
@@ -73,6 +79,8 @@ export const InputBase: FunctionalComponent<InputBaseProps> = ({
   start,
   end,
 }) => {
+  const PrefixedTagNames = getPrefixedTagNames(host);
+
   return (
     <div class="root">
       <Label
@@ -81,14 +89,17 @@ export const InputBase: FunctionalComponent<InputBaseProps> = ({
         description={description}
         htmlFor={id}
         isRequired={required}
+        isLoading={loading}
         isDisabled={disabled}
       />
       <div class="wrapper">
         <slot name="start" />
         {start}
         <input
-          aria-describedby={`${descriptionId} ${messageId}`}
+          aria-describedby={loading ? loadingId : `${descriptionId} ${messageId}`}
           aria-invalid={state === 'error' ? 'true' : null}
+          aria-disabled={disabled || loading ? 'true' : null}
+          aria-readonly={readOnly ? 'true' : null}
           id={id}
           ref={refElement}
           onInput={onInput}
@@ -112,8 +123,10 @@ export const InputBase: FunctionalComponent<InputBaseProps> = ({
         />
         {end}
         <slot name="end" />
+        {loading && <PrefixedTagNames.pSpinner class="spinner" size="inherit" theme={theme} aria-hidden="true" />}
       </div>
       <StateMessage state={state} message={message} theme={theme} host={host} />
+      <LoadingMessage loading={loading} initialLoading={initialLoading} />
     </div>
   );
 };