+++ packages/components/src/components/input-password/input-password.tsx
@@ -38,6 +38,7 @@ const propTypes: PropTypes<typeof InputPassword> = {
   name: AllowedTypes.string,
   value: AllowedTypes.string,
   required: AllowedTypes.boolean,
+  loading: AllowedTypes.boolean,
   disabled: AllowedTypes.boolean,
   maxLength: AllowedTypes.number,
   minLength: AllowedTypes.number,
@@ -108,6 +109,9 @@ export class InputPassword {
   /** Marks the password input as required. */
   @Prop() public required?: boolean = false;
 
+  /** @experimental Shows a loading indicator. */
+  @Prop() public loading?: boolean = false;
+
   /** The validation state. */
   @Prop() public state?: InputPasswordState = 'none';
 
@@ -136,6 +140,7 @@ export class InputPassword {
 
   @State() private showPassword = false;
 
+  private initialLoading: boolean = false;
   private inputElement: HTMLInputElement;
   private defaultValue: string;
 
@@ -144,13 +149,17 @@ export class InputPassword {
     this.internals?.setFormValue(newValue);
   }
 
+  public connectedCallback(): void {
+    this.initialLoading = this.loading;
+  }
+
   public componentWillLoad(): void {
     this.defaultValue = this.value;
+    this.initialLoading = this.loading;
   }
 
   public formResetCallback(): void {
-    this.internals?.setFormValue(this.defaultValue);
-    this.value = this.defaultValue;
+    this.value = this.defaultValue; // triggers value watcher
   }
 
   public formDisabledCallback(disabled: boolean): void {
@@ -173,6 +182,12 @@ export class InputPassword {
     this.internals?.setValidity(this.inputElement.validity, this.inputElement.validationMessage, this.inputElement);
   }
 
+  public componentWillUpdate(): void {
+    if (this.loading) {
+      this.initialLoading = true;
+    }
+  }
+
   public render(): JSX.Element {
     validateProps(this, propTypes);
 
@@ -180,6 +195,7 @@ export class InputPassword {
       this.host,
       getComponentCss,
       this.disabled,
+      this.loading,
       this.hideLabel,
       this.state,
       this.toggle,
@@ -214,6 +230,8 @@ export class InputPassword {
         state={this.state}
         message={this.message}
         theme={this.theme}
+        loading={this.loading}
+        initialLoading={this.initialLoading}
         end={
           this.toggle && (
             <PrefixedTagNames.pButtonPure