+++ packages/components/CHANGELOG.md
@@ -18,6 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 - `Popover`: support for custom slotted button
   ([#3861](https://github.com/porsche-design-system/porsche-design-system/pull/3861))
+- `Input Search`: ([#3874](https://github.com/porsche-design-system/porsche-design-system/pull/3874))
+- `Input Number`, `Input Password`: added loading state
+  ([#3874](https://github.com/porsche-design-system/porsche-design-system/pull/3874))
 
 ## [3.28.0] - 2025-06-02